% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_exons}
\alias{assert_valid_exons}
\title{Validate exon annotation data frame}
\usage{
assert_valid_exons(exons)
}
\arguments{
\item{exons}{exon annotation data.frame}
}
\description{
Validate exon annotation data frame
}
\keyword{internal}
