% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_genomic_coords}
\alias{assert_valid_genomic_coords}
\title{Assert that file paths are readable}
\usage{
assert_valid_genomic_coords(chr, start, end, allow_equal = FALSE)
}
\arguments{
\item{x}{A character vector of file paths to check for existence}
}
\value{
Nothing if all files exist, otherwise throws an error
}
\description{
This function checks whether all provided file paths exist and are readable.
If any file paths do not exist, it throws an informative error message.
}
\keyword{internal}
