% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_samples}
\alias{assert_valid_samples}
\title{Validate sample annotation data frame}
\usage{
assert_valid_samples(
  samples,
  required_cols = c("sample", "group"),
  context = "sample annotation"
)
}
\arguments{
\item{samples}{sample annotation data.frame}

\item{required_cols}{required column names}

\item{context}{description of where this is being used for error messages}
}
\description{
Validate sample annotation data frame
}
\keyword{internal}
