% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_reads.R
\name{cluster_reads}
\alias{cluster_reads}
\title{Cluster reads based on methylation}
\usage{
cluster_reads(x, chr, start, end, min_pts = 5)
}
\arguments{
\item{x}{a ModBamResult object.}

\item{chr}{the chromosome name where to find the region.}

\item{start}{the start position of the region.}

\item{end}{the end position of the region.}

\item{min_pts}{the minimum number of points needed to form a cluster (default = 5).}
}
\value{
A tibble with information about each read's cluster assignment and read statistics.
}
\description{
Cluster reads based on methylation
}
