% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_methy_to_dss.R
\name{convert_methy_to_dss}
\alias{convert_methy_to_dss}
\title{Convert methylation file to dss format}
\usage{
convert_methy_to_dss(methy, out_folder)
}
\arguments{
\item{methy}{the path to the methylation tabix file.}

\item{out_folder}{the folder in which to store the per sample files.}
}
\value{
data.frame of sample names and file paths
}
\description{
Convert a file in NanoMethViz's methylation format to a collection of files
in the format required by DSS::makeBSseqData()
}
