% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exons_to_genes.R
\name{exons_to_genes}
\alias{exons_to_genes}
\title{Convert exon annotation to genes}
\usage{
exons_to_genes(x)
}
\arguments{
\item{x}{the exon level annotation containing columns "gene_id", "chr",
"strand" and "symbol".}
}
\value{
the gene level annotation where each gene is taken to span the
earliest start position and latest end position of its exons.
}
\description{
Convert exon annotation to genes
}
\examples{
nmr <- load_example_nanomethresult()
exons_to_genes(NanoMethViz::exons(nmr))
}
