% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cgi.R
\name{get_cgi}
\alias{get_cgi}
\title{Get CpG islands annotations}
\value{
tibble (data.frame) object containing CpG islands annotation.
}
\description{
Helper functions are provided for obtaining CpG islands annotations from UCSC table browser.
}
