% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons.R
\name{get_example_exons_mus_musculus}
\alias{get_example_exons_mus_musculus}
\title{Get example exon annotations for mus musculus (mm10)}
\usage{
get_example_exons_mus_musculus()
}
\value{
data.frame containing exons
}
\description{
This is a small subset of the exons returned by
\code{get_exons_mus_musculus()} for demonstrative purposes. It contains
the exons for the genes Brca1, Brca2, Impact, Meg3, Peg3 and Xist.
}
\examples{
example_exons <- get_example_exons_mus_musculus()

}
