% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons.R
\name{get_exons_hg19}
\alias{get_exons_hg19}
\title{Get exon annotations for Homo sapiens (hg19)}
\usage{
get_exons_hg19()
}
\value{
data.frame containing exons
}
\description{
Get exon annotations for Homo sapiens (hg19)
}
\examples{
hg19_exons <- get_exons_hg19()

}
