% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons.R
\name{get_exons_hg38}
\alias{get_exons_hg38}
\title{Get exon annotations for Homo sapiens (hg38)}
\usage{
get_exons_hg38()
}
\value{
data.frame containing exons
}
\description{
Get exon annotations for Homo sapiens (hg38)
}
\examples{
hg38_exons <- get_exons_hg38()

}
