% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons.R
\name{get_exons_homo_sapiens}
\alias{get_exons_homo_sapiens}
\title{Get exon annotations for homo sapiens}
\usage{
get_exons_homo_sapiens()
}
\value{
data.frame containing exons
}
\description{
Get exon annotations for homo sapiens
}
\examples{
h_sapiens_exons <- get_exons_homo_sapiens()

}
