% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons.R
\name{get_exons_mm10}
\alias{get_exons_mm10}
\title{Get exon annotations for Mus musculus (mm10)}
\usage{
get_exons_mm10()
}
\value{
data.frame containing exons
}
\description{
Get exon annotations for Mus musculus (mm10)
}
\examples{
mm10_exons <- get_exons_mm10()

}
