% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons.R
\name{get_exons_mus_musculus}
\alias{get_exons_mus_musculus}
\title{Get exon annotations for mus musculus}
\usage{
get_exons_mus_musculus()
}
\value{
data.frame containing exons
}
\description{
Get exon annotations for mus musculus
}
