% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_example_modbamresult}
\alias{load_example_modbamresult}
\title{Load an example ModBamResult object}
\usage{
load_example_modbamresult()
}
\value{
a ModBamResult object
}
\description{
Load an example ModBamResult object for demonstration of plotting
functions. Run \code{load_example_modbamresult} without the function call to
see how the object is constructed.
}
\examples{
mbr <- load_example_modbamresult()
}
