% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_example_nanomethresult}
\alias{load_example_nanomethresult}
\title{Load an example NanoMethResult object}
\usage{
load_example_nanomethresult()
}
\value{
a NanoMethResult object
}
\description{
Load an example NanoMethResult object for demonstration of plotting
functions. Run \code{load_example_nanomethresult} without the function call to
see how the object is constructed.
}
\examples{
nmr <- load_example_nanomethresult()
}
