% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{methy}
\alias{methy}
\title{Get methylation data}
\usage{
methy(object)
}
\arguments{
\item{object}{the object.}
}
\value{
the path to the methylation data.
}
\description{
Get methylation data
}
\examples{
showMethods("methy")

}
\keyword{internal}
