% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_specs.R
\name{methy_col_names}
\alias{methy_col_names}
\title{Column names for methylation data}
\usage{
methy_col_names()
}
\value{
column names for methylation data
}
\description{
Column names for methylation data
}
\examples{
methy_col_names()
}
