% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_methy.R
\name{methy_data_cols}
\alias{methy_data_cols}
\title{Column names for methylation data}
\usage{
methy_data_cols()
}
\value{
column names for methylation data
}
\description{
Column names for methylation data
}
\examples{
methy_data_cols()
}
