% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModBamResults.R
\name{ModBamFiles}
\alias{ModBamFiles}
\title{Constructor for a ModBamFiles object}
\usage{
ModBamFiles(samples, paths)
}
\arguments{
\item{samples}{A character vector with the names of the samples.}

\item{paths}{A character vector with the file paths for the BAM files.}
}
\value{
A ModBamFiles object with the sample and path information.
}
\description{
This function creates a ModBamFiles object containing information about the
samples and file paths.
}
