% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoMethResults.R
\name{NanoMethResult}
\alias{NanoMethResult}
\title{Nanopore Methylation Result Constructor}
\usage{
NanoMethResult(methy, samples, exons = NULL)
}
\arguments{
\item{methy}{the path to methylation database (sql or indexed tabix).}

\item{samples}{the data.frame of sample annotation containg at least columns
sample and group.}

\item{exons}{the data.frame of exon information containing at least columns
gene_id, chr, strand, start, end, transcript_id and symbol.}
}
\value{
a NanoMethResult object to be used with plotting functions
}
\description{
Nanopore Methylation Result Constructor
}
