% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agg_regions.R
\name{plot_agg_regions_sample_grouped}
\alias{plot_agg_regions_sample_grouped}
\title{Plot aggregate regions with grouped samples}
\usage{
plot_agg_regions_sample_grouped(
  x,
  regions,
  flank = 2000,
  stranded = TRUE,
  span = 0.05,
  palette = ggplot2::scale_colour_brewer(palette = "Set1")
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{regions}{a table of regions or GRanges, or a list of such objects. The
table of regions must contain chr, strand, start and end columns.}

\item{flank}{the number of flanking bases to add to each side of each region.}

\item{stranded}{TRUE if negative strand features should have coordinates
flipped to reflect features like transcription start sites.}

\item{span}{the span for loess smoothing.}

\item{palette}{the colour palette used for samples.}
}
\value{
a ggplot object containing the aggregate methylation trend.
}
\description{
Plot aggregate regions with grouped samples
}
\examples{
nmr <- load_example_nanomethresult()
gene_anno <- exons_to_genes(NanoMethViz::exons(nmr))
plot_agg_regions_sample_grouped(nmr, gene_anno)

}
