% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate_regions.R
\name{plot_aggregate_regions}
\alias{plot_aggregate_regions}
\title{Plot aggregate regions}
\usage{
plot_aggregate_regions(
  x,
  regions,
  groups = NULL,
  flank = 2000,
  stranded = TRUE
)
}
\arguments{
\item{x}{the NanoMethResults object.}

\item{regions}{a table of regions or GRanges, or a list of such objects. The table of regions must contain chr, start and end columns.}

\item{groups}{if 'features' is a list, a vector of characters of the same length as the list containing names for each member.}

\item{flank}{the number of flanking bases to add to each side of each region.}
}
\value{

}
\description{
Plot aggregate regions
}
