% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grange.R
\name{plot_grange}
\alias{plot_grange}
\title{Plot GRanges}
\usage{
plot_grange(x, grange, anno_regions = NULL, spaghetti = FALSE, span = NULL)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{grange}{the GRanges object with one entry.}

\item{anno_regions}{the data.frame of regions to be annotated}

\item{spaghetti}{whether or not individual reads should be shown.}

\item{span}{the span for loess smoothing.}
}
\description{
Plot GRanges
}
