% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix_utils.R
\name{raw_methy_to_tabix}
\alias{raw_methy_to_tabix}
\title{Convert methylation file to tabix format}
\usage{
raw_methy_to_tabix(x)
}
\arguments{
\item{x}{the path to the sorted methylation file}
}
\value{
invisibly returns the path to the tabix file
}
\description{
Convert methylation file to tabix format
}
\examples{
methy_calls <- system.file(package = "NanoMethViz",
    c("sample1_nanopolish.tsv.gz", "sample2_nanopolish.tsv.gz"))
temp_file <- tempfile()
convert_methy_format(methy_calls, temp_file)
sort_methy_file(temp_file)

raw_methy_to_tabix(temp_file)
}
\keyword{internal}
