% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoMethResults.R
\name{samples}
\alias{samples}
\title{Get sample annotation}
\usage{
samples(object)
}
\arguments{
\item{object}{the object.}
}
\value{
data.frame of sample annotation.
}
\description{
Get sample annotation
}
