% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix_utils.R
\name{sort_methy_file}
\alias{sort_methy_file}
\title{Sort methylation file}
\usage{
sort_methy_file(x)
}
\arguments{
\item{x}{the path to the methylation file to sort}
}
\value{
invisibly returns path of sorted file
}
\description{
Sort methylation file
}
\examples{
methy_calls <- system.file(package = "NanoMethViz",
    c("sample1_nanopolish.tsv.gz", "sample2_nanopolish.tsv.gz"))
temp_file <- tempfile()
convert_methy_format(methy_calls, temp_file)

sort_methy_file(temp_file)
}
\keyword{internal}
