% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_format.R
\name{convert_methy_format}
\alias{convert_methy_format}
\title{Convert methylation calls to NanoMethViz format}
\usage{
convert_methy_format(
  input_files,
  output_file,
  samples = fs::path_ext_remove(fs::path_file(input_files))
)
}
\arguments{
\item{input_files}{the files to convert}

\item{output_file}{the output file to write results to}

\item{samples}{the names of samples corresponding to each file}
}
\value{
invisibly returns the output file path, creates a tabix file (.bgz)
and its index (.bgz.tbi)
}
\description{
Convert methylation calls to NanoMethViz format
}
\examples{
methy_calls <- system.file(package = "NanoMethViz",
    c("sample1_nanopolish.tsv.gz", "sample2_nanopolish.tsv.gz"))
temp_file <- tempfile()

convert_methy_format(methy_calls, temp_file)
}
\keyword{internal}
