% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_methy_format.R
\name{convert_methy_format}
\alias{convert_methy_format}
\title{Convert methylation calls to NanoMethViz format}
\usage{
convert_methy_format(
  input_files,
  output_file,
  samples = fs::path_ext_remove(fs::path_file(input_files)),
  verbose = TRUE
)
}
\arguments{
\item{input_files}{the files to convert}

\item{output_file}{the output file to write results to (must end in .bgz)}

\item{samples}{the names of samples corresponding to each file}

\item{verbose}{TRUE if progress messages are to be printed}
}
\value{
invisibly returns the output file path, creates a tabix file (.bgz)
and its index (.bgz.tbi)
}
\description{
Convert methylation calls to NanoMethViz format
}
\keyword{internal}
