% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{exons}
\alias{exons}
\title{Get exon annotation}
\usage{
exons(object)
}
\arguments{
\item{object}{the object.}
}
\value{
the exon annotation.
}
\description{
Get exon annotation
}
\examples{
showMethods("exons")

}
\keyword{internal}
