% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons.R
\name{get_exons}
\alias{get_exons}
\alias{get_exons_mm10}
\alias{get_exons_hg19}
\alias{get_exons_hg38}
\title{Get exon annotations}
\usage{
get_exons_mm10()

get_exons_hg19()

get_exons_hg38()
}
\value{
data.frame containing exons
}
\description{
Helper functions are provided for obtaining exon annotations from relevant
TxDb packages on Bioconductor for the construction of NanoMethResults
objects.
}
\examples{
mm10_exons <- get_exons_mm10()

hg19_exons <- get_exons_hg19()

hg38_exons <- get_exons_hg38()

}
