% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_example_nanomethresult}
\alias{load_example_nanomethresult}
\title{Load an example NanoMethResult object}
\usage{
load_example_nanomethresult()
}
\value{
a NanoMethResults object
}
\description{
Load an example NanoMethResult object
}
\examples{
nmr <- load_example_nanomethresult()
}
