% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_methy_to_dss.R
\name{methy_to_bsseq}
\alias{methy_to_bsseq}
\title{Create BSSeq object from methylation tabix file}
\usage{
methy_to_bsseq(methy, out_folder = tempdir())
}
\arguments{
\item{methy}{the path to the methylation tabix file.}

\item{out_folder}{the folder to store intermediate files. One file is created
for each sample and contains columns "chr", "pos", "total" and
"methylated".}
}
\value{
a BSSeq object.
}
\description{
Create BSSeq object from methylation tabix file
}
