% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModBamResults.R
\docType{class}
\name{ModBamFiles-class}
\alias{ModBamFiles-class}
\title{ModBamFiles class}
\description{
This is a class for holding information about modbam files. It is a
data.frame containing information about samples and paths to modbam files.
}
