% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModBamResults.R
\docType{methods}
\name{ModBamFiles}
\alias{ModBamFiles}
\alias{show,ModBamFiles-method}
\title{Constructor for a ModBamFiles object}
\usage{
ModBamFiles(samples, paths)

\S4method{show}{ModBamFiles}(object)
}
\arguments{
\item{samples}{a character vector with the names of the samples.}

\item{paths}{a character vector with the file paths for the BAM files.}

\item{object}{a ModBamFiles object.}
}
\value{
A ModBamFiles object with the sample and path information.
}
\description{
This function creates a ModBamFiles object containing information about the
samples and file paths.
}
