% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModBamResults.R
\docType{class}
\name{ModBamResult-class}
\alias{ModBamResult-class}
\title{Modbam methylation results}
\value{
a NanoMethResult object to be used with plotting functions
}
\description{
(Experimental) A ModBamResult object stores data used for NanoMethViz
#'visualisation. It contains stores a path to the methylation data, sample
information and optional exon information. The object is constructed using
the NanoMethResult() constructor function described in "Usage".
}
\section{Slots}{

\describe{
\item{\code{methy}}{a ModBamFiles data.frame specifying the samples and paths to bam
files.}

\item{\code{samples}}{the data.frame of sample annotation containing at least columns
sample and group.}

\item{\code{exons}}{the data.frame of exon information containing at least columns
gene_id, chr, strand, start, end, transcript_id and symbol.}
}}

