% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoMethResults.R
\docType{class}
\name{NanoMethResult-class}
\alias{NanoMethResult-class}
\alias{NanoMethResult}
\alias{methy,NanoMethResult-method}
\alias{samples,NanoMethResult-method}
\alias{exons,NanoMethResult-method}
\title{Nanopore Methylation Result}
\usage{
NanoMethResult(methy, samples, exons = NULL)

\S4method{methy}{NanoMethResult}(object)

\S4method{samples}{NanoMethResult}(object)

\S4method{exons}{NanoMethResult}(object)
}
\arguments{
\item{methy}{the path to the methylation tabix file.}

\item{samples}{the data.frame of sample annotation containg at least columns
sample and group.}

\item{exons}{(optional) the data.frame of exon information containing at least columns
gene_id, chr, strand, start, end, transcript_id and symbol.}

\item{object}{the NanoMethResult object.}
}
\value{
a NanoMethResult object to be used with plotting functions

the path to the methylation data.

the sample annotation.

the exon annotation.
}
\description{
A NanoMethResult object stores data used for NanoMethViz visualisation. It
contains stores a path to the methylation data, sample information and
optional exon information. The object is constructed using the
NanoMethResult() constructor function described in "Usage".
}
\section{Functions}{
\itemize{
\item \code{NanoMethResult}: Constructor

\item \code{methy,NanoMethResult-method}: methylation data path getter.

\item \code{samples,NanoMethResult-method}: sample annotation getter.

\item \code{exons,NanoMethResult-method}: exon annotation getter.
}}

\section{Slots}{

\describe{
\item{\code{methy}}{the path to the methylation tabix file.}

\item{\code{samples}}{the data.frame of sample annotation containg at least columns
sample and group.}

\item{\code{exons}}{the data.frame of exon information containing at least columns
gene_id, chr, strand, start, end, transcript_id and symbol.}
}}

\examples{
x <- load_example_nanomethresult()
methy(x)

}
