% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoMethResults.R
\docType{class}
\name{NanoMethResult-class}
\alias{NanoMethResult-class}
\title{Nanopore Methylation Result}
\value{
a NanoMethResult object to be used with plotting functions
}
\description{
Nanopore Methylation Result
}
\section{Slots}{

\describe{
\item{\code{methy}}{the path to methylation database (sql or indexed tabix).}

\item{\code{samples}}{the data.frame of sample annotation containg at least columns
sample and group.}

\item{\code{exons}}{the data.frame of exon information containing at least columns
gene_id, chr, strand, start, end, transcript_id and symbol.}
}}

