% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoMethViz-package.R
\docType{package}
\name{NanoMethViz-package}
\alias{NanoMethViz}
\alias{NanoMethViz-package}
\title{NanoMethViz: Visualise methlation data from Oxford Nanopore sequencing}
\description{
NanoMethViz is a toolkit for visualising methylation data from 
    Oxford Nanopore sequencing. It can be used to explore methylation patterns
    from reads derived from Oxford Nanopore direct DNA sequencing with
    methylation called by callers including nanopolish, f5c and megalodon. The
    plots in this package allow the visualisation of methylation profiles
    aggregated over experimental groups and across classes of genomic features.
}
\details{
The main plotting functions in this package are \code{\link[=plot_gene]{plot_gene()}} and
\code{\link[=plot_region]{plot_region()}}.
\itemize{
\item See \code{vignette("ImportingData", package = "NanoMethViz")} for how to import data from Nanopolish and f5c.
\item See \code{vignette("Introduction", package = "NanoMethViz")} for how to create visualisations using this package.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shians/NanoMethViz}
  \item Report bugs at \url{https://github.com/Shians/NanoMethViz/issues}
}

}
\author{
\strong{Maintainer}: Shian Su \email{su.s@wehi.edu.au}

}
\keyword{internal}
