% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agg_genes.R
\name{plot_agg_genes}
\alias{plot_agg_genes}
\title{Plot gene aggregate plot}
\usage{
plot_agg_genes(
  x,
  genes = NULL,
  binary_threshold = 0.5,
  group_col = NULL,
  flank = 2000,
  stranded = TRUE,
  span = 0.05,
  palette = ggplot2::scale_colour_brewer(palette = "Set1")
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{genes}{a character vector of genes to include in aggregate plot, if NULL then all genes are used.}

\item{binary_threshold}{the modification probability such that calls with
modification probability above the threshold are considered methylated, and
those with probability equal or below are considered unmethylated.}

\item{group_col}{the column to group aggregated trends by. This column can
be in from the regions table or samples(x).}

\item{flank}{the number of flanking bases to add to each side of each region.}

\item{stranded}{TRUE if negative strand features should have coordinates
flipped to reflect features like transcription start sites.}

\item{span}{the span for loess smoothing.}

\item{palette}{the ggplot colour palette used for groups.}
}
\value{
a ggplot object containing the aggregate methylation trend of genes.
}
\description{
Plot gene aggregate plot
}
\examples{
nmr <- load_example_nanomethresult()
plot_agg_genes(nmr)

}
