% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agg_regions.R
\name{plot_agg_regions}
\alias{plot_agg_regions}
\title{Plot aggregate regions}
\usage{
plot_agg_regions(
  x,
  regions,
  groups_feature = NULL,
  flank = 2000,
  stranded = TRUE,
  span = 0.05
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{regions}{a table of regions or GRanges, or a list of such objects. The
table of regions must contain chr, start and end columns.}

\item{groups_feature}{if 'features' is a list, a vector of characters of the same
length as the list containing names for each member.}

\item{flank}{the number of flanking bases to add to each side of each region.}

\item{stranded}{TRUE if negative strand features should have coordinates
flipped to reflect features like transcription start sites.}

\item{span}{the span for loess smoothing.}
}
\value{
a ggplot object containing the aggregate methylation trend.
}
\description{
Plot aggregate regions
}
\examples{
nmr <- load_example_nanomethresult()

plot_agg_regions(nmr, NanoMethViz::exons(nmr))

}
