% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate_regions.R
\name{plot_agg_regions_sample_grouped}
\alias{plot_agg_regions_sample_grouped}
\title{Plot aggregate regions with grouped samples}
\usage{
plot_agg_regions_sample_grouped(
  x,
  regions,
  flank = 2000,
  stranded = TRUE,
  span = 0.05
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{regions}{a table of regions or GRanges.}

\item{flank}{the number of flanking bases to add to each side of each region.}

\item{span}{the span for loess smoothing.}
}
\value{

}
\description{
Plot aggregate regions with grouped samples
}
