% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate_regions.R
\name{plot_aggregate_regions}
\alias{plot_aggregate_regions}
\title{Plot aggregate regions}
\usage{
plot_aggregate_regions(x, features, groups = NULL, flank = 2000)
}
\arguments{
\item{x}{the NanoMethResults object.}

\item{features}{a table of regions or GRanges, or a list of such objects. The table of regions must contain chr, start and end columns.}

\item{groups}{if 'features' is a list, a vector of characters of the same length as the list containing names for each member.}
}
\value{

}
\description{
Plot aggregate regions
}
