% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R, R/plot_methylation_heatmap.R
\name{plot_gene}
\alias{plot_gene}
\alias{plot_gene,NanoMethResult,character-method}
\alias{plot_gene_heatmap,NanoMethResult,character-method}
\title{Plot gene}
\usage{
plot_gene(x, gene, ...)

\S4method{plot_gene}{NanoMethResult,character}(
  x,
  gene,
  window_prop = 0.3,
  anno_regions = NULL,
  spaghetti = FALSE,
  span = NULL,
  gene_anno = TRUE
)

\S4method{plot_gene_heatmap}{NanoMethResult,character}(
  x,
  gene,
  window_prop = 0.3,
  pos_style = c("compact", "to_scale")
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{gene}{the gene symbol for the gene to plot.}

\item{...}{additional arguments}

\item{window_prop}{the size of flanking region to plot. Can be a vector of two
values for left and right window size. Values indicate proportion of gene
length.}

\item{anno_regions}{the data.frame of regions to annotate.}

\item{spaghetti}{whether or not individual reads should be shown.}

\item{span}{the span for loess smoothing.}

\item{gene_anno}{whether or not gene annotation tracks are plotted.}

\item{pos_style}{the style for plotting the base positions along the x-axis.
Defaults to "to_scale", plotting (potentially) overlapping squares
along the genomic position to scale. The "compact" options plots only the
positions with measured modification.}
}
\value{
a patchwork plot containing the methylation profile in the specified
region.

a patchwork plot containing the methylation profile in the specified
region.

a ggplot plot containing the heatmap.
}
\description{
Plot gene
}
\examples{
nmr <- load_example_nanomethresult()
plot_gene(nmr, "Peg3")

nmr <- load_example_nanomethresult()
plot_gene(nmr, "Peg3")

nmr <- load_example_nanomethresult()
plot_gene(nmr, "Peg3")

}
