% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot gene}
\usage{
plot_gene(x, gene, anno_regions = NULL, spaghetti = FALSE)
}
\arguments{
\item{x}{the NanoMethResults object}

\item{gene}{the gene symbol for the gene to plot}

\item{anno_regions}{the data.frame of regions to be annotated}

\item{spaghetti}{whether spaghettis should be drawn}
}
\value{

}
\description{
Plot gene
}
