% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\alias{plot_gene,NanoMethResult,character-method}
\title{Plot gene}
\usage{
plot_gene(x, gene, ...)

\S4method{plot_gene}{NanoMethResult,character}(
  x,
  gene,
  window = 0.3,
  anno_regions = NULL,
  spaghetti = FALSE,
  span = NULL
)
}
\arguments{
\item{x}{the NanoMethResults object.}

\item{gene}{the gene symbol for the gene to plot.}

\item{...}{additional arguments}

\item{window}{the size of flanking region to plot. Can be a vector of two
values for left and right window size. Values indicate proportion of gene
length.}

\item{anno_regions}{the data.frame of regions to annotate.}

\item{spaghetti}{whether or not individual reads should be shown.}

\item{span}{the span for loess smoothing.}
}
\value{
None
}
\description{
Plot gene
}
