% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\alias{plot_gene,NanoMethResult,character-method}
\alias{plot_gene,ModBamResult,character-method}
\title{Plot gene methylation}
\usage{
plot_gene(x, gene, ...)

\S4method{plot_gene}{NanoMethResult,character}(
  x,
  gene,
  window_prop = 0.3,
  anno_regions = NULL,
  binary_threshold = NULL,
  avg_method = c("mean", "median"),
  spaghetti = FALSE,
  heatmap = FALSE,
  heatmap_subsample = 50,
  span = NULL,
  gene_anno = TRUE,
  palette = ggplot2::scale_colour_brewer(palette = "Set1"),
  line_size = 2
)

\S4method{plot_gene}{ModBamResult,character}(
  x,
  gene,
  window_prop = 0.3,
  anno_regions = NULL,
  binary_threshold = NULL,
  avg_method = c("mean", "median"),
  spaghetti = FALSE,
  heatmap = FALSE,
  heatmap_subsample = 50,
  span = NULL,
  gene_anno = TRUE,
  palette = ggplot2::scale_colour_brewer(palette = "Set1"),
  line_size = 2
)
}
\arguments{
\item{x}{the NanoMethResult or ModBamResult object.}

\item{gene}{the gene symbol for the gene to plot.}

\item{window_prop}{the size of flanking region to plot. Can be a vector of two
values for left and right window size. Values indicate proportion of gene
length.}

\item{anno_regions}{the data.frame of regions to be annotated.}

\item{binary_threshold}{the modification probability such that calls with
modification probability above the threshold are set to 1 and probabilities
equal to or below the threshold are set to 0.}

\item{avg_method}{the average method for pre-smoothing at each genomic position.
Data is pre-smoothed at each genomic position before the smoothed aggregate line
is generated for performance reasons. The default is "mean" which corresponds
to the average methylation fraction. The alternative "median" option is
closer to an average within the more common methylation state.}

\item{spaghetti}{whether or not individual reads should be shown.}

\item{heatmap}{whether or not read-methylation heatmap should be shown.}

\item{heatmap_subsample}{how many packed rows of reads to subsample to.}

\item{span}{the span for loess smoothing.}

\item{palette}{the ggplot colour palette used for groups.}

\item{line_size}{the size of the lines.}
}
\value{
a patchwork plot containing the methylation profile in the specified
region.
}
\description{
Plot the methylation of a gene symbol specified within the exon(x) slot.
}
\section{Functions}{
\itemize{
\item \code{plot_gene(x = ModBamResult, gene = character)}: S4 method for ModBamResult

}}
\examples{
nmr <- load_example_nanomethresult()
plot_gene(nmr, "Peg3")

}
