% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\alias{plot_gene,NanoMethResult,character-method}
\title{Plot gene}
\usage{
plot_gene(x, gene, ...)

\S4method{plot_gene}{NanoMethResult,character}(
  x,
  gene,
  window_prop = 0.3,
  anno_regions = NULL,
  binary_threshold = NULL,
  avg_method = c("median", "mean"),
  spaghetti = FALSE,
  heatmap = FALSE,
  span = NULL,
  gene_anno = TRUE
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{gene}{the gene symbol for the gene to plot.}

\item{...}{additional arguments}

\item{window_prop}{the size of flanking region to plot. Can be a vector of two
values for left and right window size. Values indicate proportion of gene
length.}

\item{anno_regions}{the data.frame of regions to annotate.}

\item{binary_threshold}{the modification probability such that calls with
modification probability above the threshold are set to 1 and probabilities
equalt to or below the threshold are set to 0.}

\item{avg_method}{the average method for pre-smoothing at each genomic position.
Data is pre-smoothed at each genomic position before the smoothed aggregate line
is generated for performance reasons. The default is "median" which produces
a smoothed line more robust to outliers, "mean" is the old default and tends
to produce lines with values dampened towards 0.5.}

\item{spaghetti}{whether or not individual reads should be shown.}

\item{heatmap}{whether or not read-methylation heatmap should be shown.}

\item{span}{the span for loess smoothing.}

\item{gene_anno}{whether or not gene annotation tracks are plotted.}
}
\value{
a patchwork plot containing the methylation profile in the specified
region.

a patchwork plot containing the methylation profile in the specified
region.
}
\description{
Plot gene
}
\examples{
nmr <- load_example_nanomethresult()
plot_gene(nmr, "Peg3")

nmr <- load_example_nanomethresult()
plot_gene(nmr, "Peg3")

}
