% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methylation_heatmap.R
\name{plot_gene_heatmap}
\alias{plot_gene_heatmap}
\title{Plot gene methylation heatmap}
\usage{
plot_gene_heatmap(x, gene, ...)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{gene}{the gene symbol for the gene to plot.}

\item{...}{additional arguments}
}
\value{
a ggplot object of the heatmap
}
\description{
Plot gene methylation heatmap
}
\examples{
nmr <- load_example_nanomethresult()
plot_gene(nmr, "Peg3")

}
