% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grange.R
\name{plot_grange}
\alias{plot_grange}
\title{Plot GRanges}
\usage{
plot_grange(
  x,
  grange,
  anno_regions = NULL,
  binary_threshold = NULL,
  spaghetti = FALSE,
  heatmap = FALSE,
  span = NULL,
  window_prop = 0.3
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{grange}{the GRanges object with one entry.}

\item{anno_regions}{the data.frame of regions to be annotated.}

\item{binary_threshold}{the modification probability such that calls with
modification probability above the threshold are set to 1 and probabilities
equalt to or below the threshold are set to 0.}

\item{spaghetti}{whether or not individual reads should be shown.}

\item{heatmap}{whether or not read-methylation heatmap should be shown.}

\item{span}{the span for loess smoothing.}

\item{window_prop}{the size of flanking region to plot. Can be a vector of two
values for left and right window size. Values indicate proportion of gene
length.}
}
\value{
a patchwork plot containing the methylation profile in the specified
region.

a patchwork plot containing the methylation profile in the specified
region.
}
\description{
Plot GRanges
}
\examples{
nmr <- load_example_nanomethresult()
plot_grange(nmr, GRanges("chr7:6703892-6730431"))

}
