% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(
  x,
  plot_dims = c(1, 2),
  labels = colnames(x),
  groups = NULL,
  legend_name = "group"
)
}
\arguments{
\item{x}{the log-methylation-ratio matrix.}

\item{plot_dims}{the numeric vector of the two dimensions to be plotted.}

\item{labels}{the character vector of labels for data points. By default uses column names of x, set to NULL to plot
points.}

\item{groups}{the character vector of groups the data points will be coloured by.}
}
\value{
ggplot object of the MDS plot.
}
\description{
Plot multi-dimensional scaling plot using algorithm of BiocSingular::runPCA(). It is recommended this be done with
the log-methylation-ratio matrix generated by bsseq_to_log_methy_ratio().
}
\examples{
nmr <- load_example_nanomethresult()
bss <- methy_to_bsseq(nmr)
lmr <- bsseq_to_log_methy_ratio(bss)
plot_pca(lmr)

}
