% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_region}
\alias{plot_region}
\title{Plot region}
\usage{
plot_region(x, chr, start, end, anno_regions = NULL, spaghetti = FALSE)
}
\arguments{
\item{x}{the NanoMethResults object}

\item{chr}{the chromosome to plot}

\item{start}{the start of the plotting region}

\item{end}{the end of the plotting region}

\item{anno_regions}{the data.frame of regions to be annotated}

\item{spaghetti}{whether spaghettis should be drawn}
}
\value{

}
\description{
Plot region
}
