% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_region}
\alias{plot_region}
\alias{plot_region,NanoMethResult,character,numeric,numeric-method}
\alias{plot_region,NanoMethResult,factor,numeric,numeric-method}
\title{Plot region}
\usage{
plot_region(x, chr, start, end, ...)

\S4method{plot_region}{NanoMethResult,character,numeric,numeric}(
  x,
  chr,
  start,
  end,
  anno_regions = NULL,
  spaghetti = FALSE,
  span = NULL
)

\S4method{plot_region}{NanoMethResult,factor,numeric,numeric}(
  x,
  chr,
  start,
  end,
  anno_regions = NULL,
  spaghetti = FALSE,
  span = NULL
)
}
\arguments{
\item{x}{the NanoMethResult object}

\item{chr}{the chromosome to plot}

\item{start}{the start of the plotting region}

\item{end}{the end of the plotting region}

\item{...}{additional arguments}

\item{anno_regions}{the data.frame of regions to be annotated}

\item{spaghetti}{whether or not individual reads should be shown.}

\item{span}{the span for loess smoothing.}
}
\value{
None
}
\description{
Plot region
}
