% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_heatmap.R
\name{plot_region_heatmap}
\alias{plot_region_heatmap}
\alias{plot_region_heatmap,NanoMethResult,factor,numeric,numeric-method}
\title{Plot gene methylation heatmap}
\usage{
plot_region_heatmap(x, chr, start, end, ...)

\S4method{plot_region_heatmap}{NanoMethResult,factor,numeric,numeric}(
  x,
  chr,
  start,
  end,
  pos_style = c("to_scale", "compact"),
  window_prop = 0.3
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{chr}{the chromosome to plot.}

\item{start}{the start of the plotting region.}

\item{end}{the end of the plotting region.}

\item{...}{additional arguments.}
}
\value{
a ggplot object of the heatmap.
}
\description{
Plot gene methylation heatmap
}
\examples{
nmr <- load_example_nanomethresult()
plot_region_heatmap(nmr, "chr7", 6703892, 6730431)

}
