% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_violin}
\alias{plot_violin}
\title{Plot boxplots for regions}
\usage{
plot_violin(
  x,
  regions,
  binary_threshold = 0.5,
  group_col = "group",
  palette = ggplot2::scale_colour_brewer(palette = "Set1")
)
}
\arguments{
\item{x}{the NanoMethResult object.}

\item{regions}{a table of regions containing at least columns chr, strand,
start and end. Any additional columns can be used for grouping.}

\item{binary_threshold}{the modification probability such that calls with
modification probability above the threshold are considered methylated, and
those with probability equal or below are considered unmethylated.}

\item{group_col}{the column to group aggregated trends by. This column can
be in from the regions table or samples(x).}

\item{palette}{the ggplot colour palette used for groups.}
}
\value{
a ggplot object containing the methylation boxplots.
}
\description{
Plot boxplots for regions
}
\examples{
nmr <- load_example_nanomethresult()
gene_anno <- exons_to_genes(NanoMethViz::exons(nmr))
plot_violin(nmr, gene_anno)
plot_violin(nmr, gene_anno, group_col = "sample")

}
