% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_methy.R
\name{query_methy}
\alias{query_methy}
\title{Query methylation data}
\usage{
query_methy(x, chr, start, end, simplify = TRUE)
}
\arguments{
\item{x}{the path to the methylation data (tabix-bgzipped)}

\item{chr}{the vector of chromosomes}

\item{start}{the vector of start positions}

\item{end}{the vector of end positions}

\item{simplify}{whether returned results should be row-concatenated}
}
\value{

}
\description{
Query methylation data
}
