% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_methy.R
\name{query_methy}
\alias{query_methy}
\title{Query methylation data}
\usage{
query_methy(
  x,
  chr,
  start,
  end,
  simplify = TRUE,
  force = FALSE,
  truncate = TRUE
)
}
\arguments{
\item{x}{the NanoMethResults object or a path to the methylation data
(tabix-bgzipped).}

\item{chr}{the vector of chromosomes}

\item{start}{the vector of start positions}

\item{end}{the vector of end positions}

\item{simplify}{whether returned results should be row-concatenated}

\item{force}{whether to force empty output when query region 'chr' does not
appear in data. Without 'force', an empty result indicates that the
requested 'chr' appears in the data but no data overlaps with requested
region, and an invalid 'chr' will cause an error.}

\item{truncate}{when querying from ModBamFiles, whether or not to truncate
returned results to only those within the specified region. Otherwise
methylation data for entire reads overlapping the region will be returned.}
}
\value{
A table containing the data within the queried regions. If simplify
is TRUE (default) then all data is contained within one table, otherwise it
is a list of tables where each element is the data from one region.
}
\description{
Query methylation data
}
\examples{
nmr <- load_example_nanomethresult()
query_methy(methy(nmr), "chr7", 6703892, 6730431)

}
