% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix_utils.R
\name{raw_methy_to_tabix}
\alias{raw_methy_to_tabix}
\title{Convert methylation file to tabix format}
\usage{
raw_methy_to_tabix(x)
}
\arguments{
\item{x}{the path to the sorted methylation file}
}
\value{
invisibly returns the path to the tabix file
}
\description{
Convert methylation file to tabix format
}
\keyword{internal}
